+++ test/functional/rpc_net.py
@@ -96,5 +96,13 @@ class NetTest(BitcoinTestFramework):
         assert_equal(peer_info[0][0]['addrbind'], peer_info[1][0]['addr'])
         assert_equal(peer_info[1][0]['addrbind'], peer_info[0][0]['addr'])
 
+        # Check that getpeerinfo can be called for a specific peer_id
+        peer_id = peer_info[0][0]['id']
+        assert_equal(peer_info[0][0]['addr'], self.nodes[0].getpeerinfo(peer_id)[0]['addr'])
+        assert_equal(peer_info[0][0]['addrbind'], self.nodes[0].getpeerinfo(peer_id)[0]['addrbind'])
+
+        # Check that getpeerinfo fails for an invalid peer_id
+        assert_raises_rpc_error(-5, "not found", self.nodes[0].getpeerinfo, 5000)
+
 if __name__ == '__main__':
     NetTest().main()