+++ doc/release-notes.md
@@ -127,6 +127,9 @@ Low-level RPC changes
   `importmulti` with hex script argument). `signrawtransaction*` also still
   works for them.
 
+- `getpeerinfo` can now take an `id` argument to only return information about the peer with
+  the specified id.
+
 ### Logging
 
 - The log timestamp format is now ISO 8601 (e.g. "2018-02-28T12:34:56Z").